AICTE/ai-question-paper-generator/requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
transformers==4.30.2
torch==2.0.1
openpyxl==3.1.2
xlsxwriter==3.1.2
pyyaml==6.0
nltk==3.8.1
spacy==3.6.0
streamlit==1.25.0
gradio==3.41.2
click==8.1.3
tqdm==4.65.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
sentence-transformers==2.2.2
fastapi==0.100.0
uvicorn==0.23.1
python-multipart==0.0.6
jinja2==3.1.2
tabulate==0.9.0
# PDF processing libraries
PyPDF2==3.0.1
pdfplumber==0.10.3
# Additional export formats
reportlab==4.0.4
# Word document processing
python-docx==0.8.11
